+++ src/main/java/pvpmode/PvPUtils.java
@@ -7,11 +7,14 @@ import java.util.function.Supplier;
 
 import cpw.mods.fml.common.eventhandler.Event;
 import net.minecraft.command.*;
+import net.minecraft.command.ICommandSender;
+import net.minecraft.entity.*;
 import net.minecraft.entity.player.*;
 import net.minecraft.item.ItemStack;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.util.*;
 import net.minecraftforge.common.MinecraftForge;
+import pvpmode.compatibility.events.EntityMasterExtractionEvent;
 import pvpmode.overrides.EnumForcedPvPMode;
 
 public class PvPUtils
@@ -262,9 +265,9 @@ public class PvPUtils
             }
         }
     }
-
-    /**
-     * Returns whether player assigned to the supplied data is currently in
+  
+    /** 
+     * Returns whether the player assigned to the supplied data is currently in
      * PvP.<br/>
      * If a PvP event occurred with this player involved, a timer starts. While
      * this timer is running, the player is considered to be involved into PvP.
@@ -287,6 +290,26 @@ public class PvPUtils
         else if (command.getCommandAliases () != null)
             return command.getCommandAliases ().contains (name);
         else return false;
+    }
+  
+    /**
+     * Returns the player that this entity is associated with, if possible.
+     */
+    public static EntityPlayerMP getMaster (Entity entity)
+    {
+        if (entity == null)
+            return null;
+
+        if (entity instanceof EntityPlayerMP)
+            return (EntityPlayerMP) entity;
+
+        if (entity instanceof IEntityOwnable)
+            return (EntityPlayerMP) ((IEntityOwnable) entity).getOwner ();
+
+        // Via this event the compatibility modules will be asked to extract the
+        // master
+        EntityMasterExtractionEvent event = new EntityMasterExtractionEvent (entity);
+        return PvPUtils.postEventAndGetResult (event, event::getMaster);
     }
 
 }